Create Gift Shop and Coupon
Chef wants to impress Chefina by giving her the maximum number of gifts possible.
Chef is in a gift shop having N items where the cost of the i th item is equal to Ai. chef has K amount of money and a 50% off discount coupon that he can use for at most one of the items he buys. If the cost of an item is equal to X, then, after applying the coupon on that item, Chef only has to pay ⌈X/2⌉(rounded up to the nearest integer) amount for that item.
Help Chef find the maximum number of items he can buy with K amount of money and a 50% discount coupon given that he can use the coupon on at most one item.
input : 
3
1 4
5
3 15
4 4 5
3 10
6 7 4
output:
1
3
2