Create implement stack with the help of queue
in this method we have used 2 queue to act as stack.
suppose we want to append 1,2,3,4 in stack so what we will do, we will take 2 empty queue as q1 and q2, initially we will append 1 in q1, and 2 in q2 and now, we will pop 1 from q1 and will append in q1, now we will Change q1 as q2 and q2 as q1, and will repeat the process again, will add 3 in q2, now will pop 2 and 1 from q1 and append in q2, so q2 will become 3 2 1, now again q1 = q2 and q2 = q1. will repeat the process 1 more time.